Implement UnknownPolicyRater using Null Object Pattern
The intent of the Null Object Pattern is to minimize that kind of null check.
Instead, we can identify the null behavior and encapsulate it in the type expected by the client code.
More often then not, such neutral logic is very simple – do nothing.
This way we no longer need to deal with special handling of null references.
https://www.baeldung.com/java-null-object-pattern